refactor: migrate bundle scripts to build:website
