+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 dist
+/page/
 test-results
 .tshy/
 .tshy-build
