+++ src/page/index.html
@@ -54,6 +54,7 @@
         </td>
       </tr>
     </table>
+    <a href="https://github.com/ghostery/urlfilter2dnr">GitHub</a>
   </section>
   <script type="module" src="./index.js"></script>
 </body>
